Base Class WIP
almost back to working

Added an update endpoint for the base data that makes use of the base class.

The Base actions are still partially broken, and will need to be fixed before merging to main.

Patch is technically more correct, also need the categories in the upgrades

fixed actions using the base class

Mechs not boats